Move levels to package
