+++ src/main/java/uib/bamboozle/levels/Level1.java
@@ -1,4 +1,4 @@
-package uib.bamboozle;
+package uib.bamboozle.levels;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Input;
@@ -7,11 +7,11 @@ import com.badlogic.gdx.physics.bullet.collision.btCollisionObject;
 import uib.bamboozle.ui.GameObject;
 import uib.bamboozle.ui.Graphics;
 
-public class Level2 extends Level {
+public class Level1 extends Level {
     private GameObject cube;
     private GameObject ball;
 
-    public Level2(Graphics graphics) {
+    public Level1(Graphics graphics) {
         super(graphics);
 
         cube = getModelFactory().get("cube", new Vector3(0f, 0f, 0f), 0f);
@@ -35,7 +35,12 @@ public class Level2 extends Level {
     }
 
     public boolean isFinished() {
-        return false;
+        if(ball != null) {
+            Vector3 pos = ball.getInstance().transform.getTranslation(new Vector3());
+            return pos.len() > 5;
+        } else {
+            return false;
+        }
     }
 
     public GameObject getCube() {
